Fix some issues with tests
